requirements.txt
absl-py==2.0.0
altgraph==0.17.4
asttokens==2.4.1
attrs==24.3.0
backcall==0.2.0
beautifulsoup4==4.12.3
bleach==6.2.0
blinker==1.9.0
cachetools==5.3.2
certifi==2023.7.22
charset-normalizer==3.3.2
click==8.2.1
colorama==0.4.6
contourpy==1.2.0
cycler==0.12.1
decorator==5.1.1
defusedxml==0.7.1
docopt==0.6.2
dominate==2.9.1
et-xmlfile==1.1.0
excel-util==0.0.4
executing==2.0.1
fastjsonschema==2.21.1
filelock==3.13.1
Flask==3.1.1
flask-cors==6.0.1
fonttools==4.44.3
fsspec==2023.10.0
gdown==5.2.0
google-api-core==2.24.0
google-api-python-client==2.156.0
google-auth==2.23.4
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.1.0
googleapis-common-protos==1.66.0
grpcio==1.59.2
httplib2==0.22.0
idna==3.4
imageio==2.34.2
intel-openmp==2021.4.0
ipython==8.12.3
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyterlab_pygments==0.3.0
kiwisolver==1.4.5
labelImg==1.8.6
lazy_loader==0.4
lpips==0.1.4
lxml==5.2.1
Markdown==3.5.1
MarkupSafe==2.1.3
matplotlib==3.8.4
matplotlib-inline==0.1.6
mistune==3.0.2
mkl==2021.4.0
MouseInfo==0.1.3
mpmath==1.3.0
mysql-connector-python==8.1.0
nbclient==0.10.2
nbconvert==7.16.4
nbformat==5.10.4
networkx==3.2.1
numpy==1.23.5
oauthlib==3.2.2
opencv-python==4.8.1.78
openpyxl==3.1.2
packaging==23.2
pandas==2.2.3
pandas-util==0.0.1
pandocfilters==1.5.1
parso==0.8.3
pefile==2023.2.7
pickleshare==0.7.5
Pillow==10.1.0
pipreqs==0.5.0
platformdirs==4.3.6
prompt-toolkit==3.0.41
proto-plus==1.25.0
protobuf==4.21.12
psutil==5.9.6
pure-eval==0.2.2
py-cpuinfo==9.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
PyAutoGUI==0.9.54
PyGetWindow==0.0.9
Pygments==2.16.1
pyinstaller==6.11.1
pyinstaller-hooks-contrib==2024.10
PyMsgBox==1.0.9
pyparsing==3.1.1
pyperclip==1.9.0
PyQt5==5.15.10
PyQt5-Qt5==5.15.2
PyQt5-sip==12.13.0
PyRect==0.2.0
PyScreeze==1.0.1
PySocks==1.7.1
python-dateutil==2.8.2
python-pptx==0.6.23
pytweening==1.2.0
pytz==2023.3.post1
pywin32==308
pywin32-ctypes==0.2.3
PyYAML==6.0.1
pyzmq==26.2.0
referencing==0.35.1
requests==2.31.0
requests-oauthlib==1.3.1
rpds-py==0.22.3
rsa==4.9
scikit-image==0.24.0
scipy==1.11.3
screeninfo==0.8.1
seaborn==0.13.2
six==1.16.0
soupsieve==2.6
stack-data==0.6.3
sympy==1.12
tbb==2021.12.0
tensorboard==2.15.1
tensorboard-data-server==0.7.2
thop==0.1.1.post2209072238
tifffile==2024.7.2
tinycss2==1.4.0
torch==2.3.0
torchvision==0.18.0
tornado==6.4.2
tqdm==4.66.1
traitlets==5.13.0
typing_extensions==4.8.0
tzdata==2023.3
ultralytics==8.2.15
uritemplate==4.1.1
urllib3==2.1.0
wcwidth==0.2.10
webencodings==0.5.1
Werkzeug==3.1.3
XlsxWriter==3.2.0
yarg==0.1.9
